requirements.txt
django==1.8.6
line-bot-sdk==2.3.0
requests==2.28.0
pyquery==2.0.0
google-generativeai